requirements.txt
pandas
python-docx
matplotlib